+++ test_runner/src/main/kotlin/ftl/http/ExecuteWithRetry.kt
@@ -3,6 +3,8 @@ package ftl.http
 import com.google.api.client.googleapis.services.json.AbstractGoogleJsonClientRequest
 import com.google.api.client.http.HttpResponseException
 import ftl.config.FtlConstants
+import ftl.util.PermissionDenied
+import ftl.util.ProjectNotFound
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.runBlocking
 import java.io.IOException
@@ -14,8 +16,7 @@ import kotlin.math.roundToInt
 fun <T> AbstractGoogleJsonClientRequest<T>.executeWithRetry(): T = withRetry { this.execute() }
 
 private inline fun <T> withRetry(crossinline block: () -> T): T = runBlocking {
-    var lastErr: IOException? = null
-
+    var lastError: IOException? = null
     repeat(4) {
         try {
             return@runBlocking block()
@@ -24,16 +25,19 @@ private inline fun <T> withRetry(crossinline block: () -> T): T = runBlocking {
             // https://github.com/Flank/flank/issues/701
             FtlConstants.bugsnag?.notify(FlankGoogleApiError(err))
 
-            lastErr = err
-            // HttpStatusCodes from google api client does not have 429 code
-            if (err is HttpResponseException && err.statusCode == 429) {
-                return@repeat
+            lastError = err
+            if (err is HttpResponseException) {
+                // we want to handle some FTL errors with special care
+                when (err.statusCode) {
+                    429 -> return@repeat
+                    403 -> throw PermissionDenied(err)
+                    404 -> throw ProjectNotFound(err)
+                }
             }
             delay(exp(it - 1.0).roundToInt().toLong())
         }
     }
-
-    throw IOException("Request failed", lastErr)
+    throw IOException("Request failed", lastError)
 }
 
 private class FlankGoogleApiError(exception: Throwable) : Error(exception)